+++ compiler/src/test/fixtures/non_inlined_fragments/output/java-client/com/example/non_inlined_fragments/EventType.java
@@ -1,0 +1,10 @@
+package com.example.non_inlined_fragments;
+
+import javax.annotation.Generated;
+
+@Generated("org.realityforge.giggle.Main")
+public enum EventType {
+  Wildfire,
+
+  Burn
+}