+++ compiler/src/test/fixtures/non_inlined_fragments/output/java-client/com/example/non_inlined_fragments/GraphQLException.java
@@ -1,0 +1,21 @@
+package com.example.non_inlined_fragments;
+
+import javax.annotation.Generated;
+
+@Generated("org.realityforge.giggle.Main")
+public class GraphQLException extends RuntimeException {
+  public GraphQLException() {
+  }
+
+  public GraphQLException(final String message) {
+    super( message );
+  }
+
+  public GraphQLException(final String message, final Throwable cause) {
+    super( message, cause );
+  }
+
+  public GraphQLException(final Throwable cause) {
+    super( cause );
+  }
+}